SD Card Folder Structure replication.
In the preference settings there is now a sd card section regrouping the sd card auto detect option and the new base folder to replicate.
If the path is defined it will copy an gcode from stl located under that folder and place it at the same location on the sd card.