Updates to Mini quick print settings
Added changes from the print profiles repo. It was also found that the nozzle setting in the machine profile was not being used. This means all of the previous gcode was using 0.4mm nozzle size for the Mini. The profiles need further testing but we could see problems with the nozzle size change.