Add a tentative sleep prevention for linux.
I wasn't able to test this, but at least it won't hurt and it's better
than nothing.
I tested the command execution and it doesn't give an error or anything
but I'm unable to get my setup to suspend so I can't test the prevention
of suspend.
This is for Linux support on issue #46