Use the right plugin configuration when running the engine
The simple settings would override the plugin config to disable all
then run the engine then reset the overrides, but the plugin config
was not accessed until the slicing engine finished, at which point
the plugin config override was already reset. this caused plugins
to be kept enabled even in simple mode.
This fixes issue #59