package.sh: add fedora
Fedora builds Cura by using ```mock```, thereby enabling it to build RPMs
for every distribution that ```mock``` has a configuration file for. In
pratice this means that Fedora can build RPMs for several versions of
Fedora, CentOS and RHEL.

Cura can be built under a regular user account, there is no need to have
root privileges. In fact, having root privileges is very much discouraged.

However, the user account under which the build is performed needs to be
a member of the 'mock' group.

Specify only 'fedora' to build RPMs for the current system.
  ./package.sh fedora

Additionally specify any number of mock configurations
(ls -1 /etc/mock/*.cfg) to build RPMs for those configurations.
  ./package.sh fedora fedora-21-x86_64.cfg fedora-21-i386.cfg ...

Since only the filename of the mock configurations is used, this
also works:
  ./package.sh fedora /etc/mock/fedora-21-x86_64.cfg ...

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>