Do not reuse an argument variable as iterator
This caused issues with the head/gantry size when printing one object
at a time. The obj variable would never be None at the end of the function
which means it wasn't updating all objects with the gantry size, only the
last object iterated on, which caused the objects to think there are no
collisions between them when printing order was being decided.
This should hopefully fix issue #104