Fix an issue where the UI state wasn't being updated
This could happen if you pause a print within 500ms from a gcode command
being executed on the serial com. The update was ignored and if we paused
we may never get another callback update so the UI would never get updated.
This could also happen when we just open the UI and we see the temperature
values appear meaning the connection was successful, but it doesn't say
Operational and the Print button is disabled for about 5 seconds (the next
temperature update).