Prevent multiple calls to afterSplashCallback
This fixes issue #133. Problem was that if splash.OnClose was called
before the afterSplashCallback was done (such as, if it's blocking
on a configWizard), then the callback would be called again.
This also fixes a segmentation fault if you cancel the initial config wizard
by making sure the callback is done outside the event thread and splash
destruction is done at the proper time and from the proper thread.