Adding script to create lulzbot profiles from print profile files
This script needs to be run from the root of cura repo, it expects
the print profiles to be in resources/quickprint/lulzbot_profiles and
it assumes a specific naming convention of the files. It will
create new directories and files for the quickprint settings based on
the files in lulzbot_profiles, according to the mappings in the script.
Any file that doesn't match will be ignored