Cancel resend if firmware asks for line 1 or more than 100 lines before current
This is because it seems when we use a heat-and-wait command (possibly),
there is a risk of getting a checksum error and resend. That could happen
for more than the 10 lines we had set previously, so setting it to 100 sounds
like a safer value.
We also check for position 1 because if the error happens in the first 100
lines, it would be missed.
We're also printing a message to the console when that happens.
Fixes T178