Better fix for T35
Reverts previous commit and fixes it properly. The issue wasn't that
wx would send a setting change event after the control was destroyed,
but rather, it would send a setting change event on the list box just
before the destroy event.
The issue was that OnSettingChange would reset the config for all of
the controls, instead of only for the control that changed, and so it
was trying to access already destroyed controls.