Remove old version of LegacyProfileReader
This shouldn't have been committed. I suspect this came in via checking out a branch where this was still on the working tree and then going to a branch where it wasn't, and it kept the files and I thought LegacySettings was the plugin I had to commit.

Contributes to issue CURA-37.