Always load CuraEngineBackend plugin first
In the same way that consolelogger is loaded firstly, the engine is loaded secondly. After that the rest of the plugins are loaded. I'd really have loved to use some sort of plugin dependency system but that is out of scope right now. Fixes all external plugins that use the backend, such as to trigger a reslice (such as PostProcessing).

Contributes to issue CURA-443.