Fix pyserial 3.0 compatibility issue
The pyserial 3.0 changed the Serial parameter writeTimeout into
write_timeout, which breaks the code. They have a fix planned for
a soon-to-be-released version, but since it's a deprecated argument
anyway, it's best not to use in order to stay future-proof, and also
in case someone installs the 3.0 version itself.
This fixes T355