Let Cura set default output file type to g-code
The default preference for last remembered output file type is set to G-code. So at the first run, it'll have g-code selected by default. In subsequent runs it will remember what the setting was, so the user can set it to something else and it will remember that.

Contributes to issue CURA-611.