Use MeshBuilder to make convex hull mesh
The convex hull mesh was made without adding colours to the vertices. This was probably causing the convex hull to appear red on some machines, if red is the default vertex colour. Using meshbuilder makes code maintenance easier and allows us to easily set the colour of the vertices. Hopefully this fixes the issue with the red convex hulls.

Contributes to issue CURA-625.