Alternative fix for NVidia graphic cards
As I upgraded my computer to a developer version of Ubuntu, I noticed that the OpenGL module, which is needed for our workaround, is currently broken.
So I browsed the web and found the reason why it is needed at all and found a alternative.
The reason for the problem is that Qt5 is dynamicly loading libGL.so instead of libGL.so.1, as the OpenGL module loads libGL.so.1.
So if you install the closed-source nvidia drivers it only creates a link from libGL.so.1 to it's binaries and the result is that PyQt5/Qt5 tries to load Mesa binaries together with NVidia binaries.
By importing the OpenGL module you preload the libGL.so.1, but this can also be done directly by using ctypes.

* Replaced the OpenGL fix with the ctypes fix
* Added a TODO