pauseAtZ plugin: Make 15mm in Z park height configurable
Just after retracting, and before parking the head, we move the Z
upwards.  Previously the minimum parking Z height was always 15mm.
The comment says:

  Change z before doing the move because the nozzle
  can hit the glass lock on the UM2

On my TAZ-5, this is not necessary.  And moving the Z axis a long
distance before parking is undesirable because Z axis moves are slow,
leading to a certain amount of ooze, etc.

So make the minimum Z parking height configurable.  We retain the
previous fixed value of 15mm as the default.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>