pauseAtZ plugin: Support re-homing X and Y before restarting
In a two-colour print, the operator has to mess about with the X
carriage to swap the filament in the print head, which can disturb the
X position.  In other applications of pauseAtZ, the operator needs to
interact with the workpiece, perhaps disturbing the Y position.

So support homing the X and/or Y axes just before resuming.

The default head park position of [150,0,] is suitable for use for
swapping filament, and moves the printbed out of the way.  So make the
default by to home X but not Y.

In both cases we travel fast to nearish the endstop (within 25mm), and
then do the home (which is slower).  This gets us going as quickly as
possible (which is helpful to avoid ooze).

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>