Only send enough commands to fill up the fw cmdbuffer again after a print/pause
The issue here is that if the user has some fun by pressing pause/resume
a lot of times, then we spam the firmware with 4 new commands every time
it resumes, but the cmdbuffer might be full, so it might end up filling
the serial ringbuffer and we end up with commands that are not sent fully.
This ensures that if a command is still being executed, only the
remaining slots in the firmware's cmdbuffer will be filled.