Queue commands even in operational mode to avoid saturating the ringbuffer
If we send manually a lot of commands even in operational mode and
the firmware is busy on a long command (G28/G29/M109/M190), we could
end up filling the ringbuffer and causing an overflow which will
cause commands to be lost. Since those commands were sent by the user,
then the firmware can't request a resend and they will therefore be lost.