Add support for M108 to cancel heatup.
If we are heating up with M109 or M190, then we can't send any new
commands until the command buffer queue in firmware is empty (M109/M190 done)
otherwise, we will fill up the buffer queue (2 more commands will be accepted after M109/M190)
and anything else we send will just end up in the serial ringbuffer and will not be read until
the M109/M190 is done.
So if we want to cancel the heatup, we need to send M108 which needs to be able to be read from
the ringbuffer, which means the command queue needs to be empty.
So we stop sending any commands after M109/M190 unless it's M108 (or until heating done) if we want
M108 to get handled.
the small delay that is caused by the firmware buffer getting empty is not important because
this happens during a heat&wait, not during a move command.
If M108 is received, it gets sent asynchronously, which stops the heatup process and
everything else that was queued is then dequeued.

Fixes T373 but needs more rigorous testing.