Update per-object settings list when a setting changes
The event triggers a refresh of the global_only property. I think it shouldn't affect the computation time too much but it feels sluggish when you change a setting. Does it feel sluggish for you too, reviewer? It has been feeling sluggish for a while so I don't know if that's due to this change.

Contributes to issue CURA-458.