Removed superfluous override from UM2+ definition
The default is line_width * 2 but this defined it as machine_nozzle_size * 2. I talked it over with Tom (who added that line) and he said it was only copied from Paul's profiles. It should've been line width. Makes more sense. It's line_width * 2 so that it can go from an outer wall to the infill without retracting.

Contributes to issue CURA-346.