Add extruders model
Hasn't been tested much because there is no display for this in the front-end yet. It needs a setting data type to go with in order to display it. This model is for the drop-down boxes to select an extruder to print, for instance, the support with.

Contributes to issues CURA-1278 and CURA-351.