Don't re-create user profile if one is loaded from files
The user profiles are also saved because they are added to the container registry. So if one exists for this extruder, don't create a new blank name, but use the pre-loaded one which contained the user settings the user had before he closed down Cura previously.

Contributes to issues CURA-340 and CURA-1278.