Mark certain types of exceptions as fatal and abort the application if they occur
We simply cannot recover properly from things like an uncaught
MemoryError since that usually means any follow up operation will also
fail. So Instead of silently ignoring it and having the application in a
broken state we properly abort.

Right now the list of fatal exceptions is a bare minimum that contains
the most prominent things we cannot recover from.