Machine head polygon with measured dimensions
The collision polygon of the machine head is rectangular, but the arm on which the nozzle hangs extends to the edge of the build plate on one side. These measurements were taken also by Calvindog717. The orientation is guessed. Usually the Y+ direction is towards the front... Let's hope it's good!