Define extruder_prime_pos_* in fdmextruder too
We need to have it in both, because it needs to be changed by the inheriting definitions (so it needs to be defined in fdmextruder) but it also needs to be defined for single-extrusion printers (so it needs to be in fdmprinter). I copied these definitions from fdmprinter.

Contributes to issue CURA-1816.