CURA-1223: 3MFReader: Doing selftest and fail on broken files
This commit adds a selftest before the "result" gets returned.
It should break on this/these functions and popup a message about using
the online repair tool.
I compared the content of both files for a long time, googled about the
format, but wasn't able to find a proper fix.
More routines will likely need to be added here, but with those I have
it is working so far very well.

Sadly it is not possible to override the default message, so two
messages will appear. Additionally, the URL of the link is not
clickable/executable from the UX. Just acting like normal text.

Contributes to CURA-1223