Scanning for CuraEngine and stop if it was not found
So far I ran 2 or 3 times into the problem that my engine was not up to
date. The bad thing here is that there is always an updated version from
my PPA on the PC. So first this commit gives the possibility to look for
CuraEngine in the $PATH directories. Before I had to copy it always
manually, eg. via "cp $(which CuraEngine) bin/CuraEngine" in my
workbench.

Second, people can get into the situation that CuraEngine is missing at
all. So before making Cura loop and try to use an executable that does
not exist, better raise an Exception here.
An additional info message tells about the location being used.

Does not contribute to any JIRA issue (I think), but makes my life
easier.