Allow extruderValue(s) to fallback to global
If a setting is not defined in the extruder stack, fall back to global. At least it still tries to get the extruder value from the correct stack first. If that stack has no value, then so be it, use global anyway.

Contributes to issue CURA-2024. Also could affect issue CURA-2006.