Allow higher speeds for diagonal wireframe moves
Since these moves are slightly diagonal in all directions, a higher speed can be achieved. Exactly how high depends on both angles, but for simplification this limit is a bit too broad, allowing speeds that the printer can't handle. The firmware should catch that case if it happens.

Contributes to issue CURA-1049.