Update Innovo Inventor JSON and STL files
Made some changes to the Inventor JSON file:
- made machine center is zero true
- added turn heated bed off to end g code
- changed top/bottom thickness from 0.30mm to 1.2mm (don’t know why I
put it to 0.3 in the first place!!)
- changed material bed temp to 0 by default

As for the STL file, it’s been updated to a much smaller file, 810kb
(down from 13mb).  I thought I updated this before but apparently not.

I hope this update can be merged soon, because I plan on launching a
makerspace with these printers and everyone will be using Cura so they
need these settings