Add Kossel Mini printer definition
Added Kossel Mini Printer definition. It is the first delta printer working on Cura; I tested it before uploading this .def.json file and its accompanying .stl file. Works but there is some mysterious padding at the sides of the "square" (?) heated bed that disallow the part but these are not areas I configured. I approximated the round shape of the bed with 8 polygons for disallowed areas. However the square checkerboard pattern appears on top of it and it is kind of annoying, hope someone would fix this.