CURA-1445: Sending bundle of serialized data
After an discussion with the developers at YM, we decided to send all data in one keyword.
So now all data is bundled into a JSON dictionary, where "" is the global_container_stack and it's containers go into the same JSON, too, by using their IDs as keywords.
More on JIRA...

Contributes to CURA-1445

Includes/Fixes #1013