T522: Use the proper profile settings from the gcode when interpreting it
Partial fix to T522. The gcode interpreter was using local profile settings
that would drasticly affect the gcode interpretation, and was ignoring
the actual profile settings that were used to generate that gcode file.
This would cause invalid results to be returned, especially in the case
of a gcode generated with the Spiralize option. This will look for the
cura profile string and load the settings from it to use in the interpretation
of the gcode.