Prevent first layer from being too thin by properly setting its size
This is part of T522, basically, going from Z position 2.0 from start
gcode to the first layer at 0.4, caused the gcode interpreter to think
that the first layer height was 1.6mm. Also, with the moarstruder, the
first layer can be thicker than 1mm, so I've set it to check for Z
heights of up to 1.5mm.