Add invisible character to satisfy gettext
This is a zero-width space. It should be invisible in the interface. It is required, however, because gettext breaks when it sees that the original text has \n and the translation doesn't. In this case, the \n is just our text though, and it is allowed to be in a different place in the translation. So I'm silencing gettext by adding a character in front of it.

Contributes to issue CURA-570.