Publish Ultimaker 3 definitions and profiles
The Ultimaker 3 is now released, so we can make this data public. These are the Ultimaker 3 definition files (with their physical properties as Cura requires them, such as build volume) as well as its extruders, print cores, quality profiles and build plate mesh. And also for the Ultimaker 3 Extended. Modifications to the material profiles will be published in the fdm_materials repository.

Contributes to issue CURA-2737.