Remove code that sets incompatible materials to invalid type
Instead use the "compatible" metadata entry like we do for nozzles.

Setting the type differently caused all code that looks for materials to
fail because "incompatible_material" is not "material". Since this is
the base file, this meant a lot of materials failed to duplicate
properly which caused crashes.

Contributes to CURA-2821