Filter materials on compatibility for non-machine-specific materials
This only listens to whether the 'base' XML profile is marked as compatible in its metadata, but since the machine doesn't have its own materials, the base is the only one that exists. This needs to be changed when we change the filtering on fdmprinter, hence the comment in the code.

Contributes to issue CURA-2837.