Compute disallowed areas per-extruder
This computes the disallowed areas of the machine, the disallowed border and the prime tower per extruder. The prime locations are not yet calculated per extruder, but I'd like to extract it to a function before I try that.

Warning: Cura will not run at this stage. Please await the next commits.

Contributes to issue CURA-2625.