Move right two clips to head disallowed areas
It turns out that the right nozzle, if lifted, can still hit the clips in very rare cases. But the left nozzle, if the right nozzle is down, will not hit the clips. So the right clips are now part of the static areas just like the switching bay and camera, but the left clips are not.

Contributes to issue CURA-2625.