Fixed errant bytearrays
These caused the system to always use the default path ("/"), occasionally leaving it in an inconsistent state.

The GUI would be able to set the path to the correct value, but the correct value would not be reflected in the GUI, even though the preferences file was occasionally correct.

Since the system used the same function to read from a preference file as it did to create the GUI object, the GUI and internal data structures would show the default path, and the preference file would show the correct value.