Using UM.Platform and correcting something minor
Since we already have UM.Platform, I think it is worth to get constancy into our code and use UM.Platform for our detection.
Additionally, I corrected the head of the file, because I think I shall mention Cura and not Uranium. I guess the plugin was moved to Cura, so this was missed.