Add quality profiles for ABAX printers
These are delivered to me by Pedro J. of ABAX 3D Tech. I modified the file names to be in line with our code standards, and I modified the link to the machine definition to coincide with the file names that were changed there to be in line with our code standards.