Add label for printer connection text to header
Also fixed the indenting of the previous two labels to use spaces instead of tabs. I knew I was going to forget putting it back to spaces after working on a different project that uses tabs.

The information is duplicated now. I'll remove the old one promptly.

Contributes to issue CURA-3161.