Added anchor_skin_shrink_distance.
Also, renamed anchor_skin_distance to anchor_skin_expand_distance.

The idea behind the shrink distance is that when the slope of the model surface
is steep, very slim skin areas are created close to the wall and if they
are expanded we end up with skin inside the infill that isn't required. So
by shrinking the skin polygons slightly first, the very slim areas are removed
before the skin is expanded. The amount to shrink defaults to half the wall
width which appears to work OK but may as well make it a setting so that it
can be tweaked if required.