Anchor last-row extruder box to left and right to stretch it
Layout.fillwidth seems to only stretch the box to full width on the first row, but without it the entire thing doesn't stretch. Leaving the width out will make the entire left column stretch so that the right column is no longer visible. It's all a bit weird, this QML stuff.

Contributes to issue CURA-3161.