Add support for WSH USB driver when not filtering USB-only
This support was included when filtering for USB only, but not the case when we are not filtering. I don't like this bit of code much since we should just define a list of paths to check and iterate over it, so that we can re-use the data and prevent syncing mistakes like this.