Add test for deserialising materials in stacks
Similar to the other tests, but this one is a bit special since the original test stacks had a material defined. Therefore it also defines a separate stack that only has a definition so that we can see if it's fine with the material being empty.

Contributes to issue CURA-3497.