Ups the limit of the maximum cost of a spool of filament
This is necessary for currencies such as Japanese Yen or South Korean Won, where prices can easily exceed 1000. This PR sets the limit to 100,000,000. If the price of a spool exceeds this amount, perhaps the user should consider moving to a more stable country.

Fixes https://github.com/Ultimaker/Cura/issues/1567