Update XmlMaterialProfile.py
added retraction_prime_speed  as a material specific value.
This way, if a material needs a specific setting, the value set in the material xml file will be used. 
Probably a good idea to have this value added in the front end as well, but for now this will do.