Optimise 0.4mm profiles for CPEP, PC and TPU
These optimisations should only remove settings that are exactly equal to the parent profile. Support interface settings are an exception, since they were defining non-existing settings in the original profiles. The stand-by temperature is also an exception, since the stand-by temperature of the material profile is set in stone because of firmware.

Contributes to issue CURA-3650.