Simplify gradual infill steps maximum formula
The case where it reaches '20' is incorrect, it should never give an error, so it should've been 999999 or something. This case is also overly complex.
Say that P is 'infill_line_distance > 0' and Q is 'spaghetti_enabled'...
Then it would be the logarithmic if (P and not Q)...
Then it would be 0 or 20 if (not (P and not Q)), which is ((not P) or Q)...
Then it would be 20 if (((not P) or Q) and not Q), which is (not P)...
And it would be 0 if (((not P) or Q) and Q), which is Q.

So therefore it can be simplified to: 20 if (not P) else 0 if Q else logarithmic. But 20 became 999999, hence the above formula.

Contributes to issue CURA-3700.